+++ CotEditor/Sources/Document.swift
@@ -683,40 +683,53 @@ final class Document: NSDocument, AdditionalDocumentPreparing, EncodingChanging 
     
     // MARK: Protocols
     
-    /// File has been modified by an external process.
     override func presentedItemDidChange() {
         
         // [caution] This method can be called from any thread.
         
-        // [caution] DO NOT invoke `super.presentedItemDidChange()` that reverts document automatically if autosavesInPlace is enable.
+        // [caution] DO NOT invoke `super.presentedItemDidChange()` that reverts document automatically if autosavesInPlace is enabled.
 //        super.presentedItemDidChange()
         
         guard
             UserDefaults.standard[.documentConflictOption] != .ignore,
             !self.isExternalUpdateAlertShown,  // don't check twice if already notified
-            let fileURL = self.fileURL
+            var fileURL = self.fileURL
         else { return }
         
-        // ignore if file's modificationDate is the same as document's modificationDate
-        let modificationDate: Date?
-        do {
-            modificationDate = try fileURL.resourceValues(forKeys: [.contentModificationDateKey]).contentModificationDate  // FILE_ACCESS
-        } catch {
-            return assertionFailure(error.localizedDescription)
-        }
-        guard let modificationDate, modificationDate != self.fileModificationDate else { return }
-        
         // check if the file content was changed from the stored file data
-        var data: Data?
+        var didChange = false
+        var modificationDate: Date?
         var error: NSError?
-        NSFileCoordinator(filePresenter: self).coordinate(readingItemAt: fileURL, options: .withoutChanges, error: &error) { newURL in
+        NSFileCoordinator(filePresenter: self).coordinate(readingItemAt: fileURL, options: .withoutChanges, error: &error) { newURL in  // FILE_ACCESS
+            var newURL = newURL
+            fileURL.removeCachedResourceValue(forKey: .contentModificationDateKey)
             do {
-                data = try Data(contentsOf: newURL, options: [.mappedIfSafe])  // FILE_ACCESS
+                // ignore if file's modificationDate is the same as document's modificationDate
+                modificationDate = try fileURL.resourceValues(forKeys: [.contentModificationDateKey]).contentModificationDate
+                guard modificationDate != self.fileModificationDate else { return }
+                
+                // check if file contents was changed from the stored file data
+                let data = try Data(contentsOf: newURL, options: [.mappedIfSafe])
+                didChange = data != self.fileData
             } catch {
-                return assertionFailure(error.localizedDescription)
+                assertionFailure(error.localizedDescription)
             }
         }
-        guard let data, data != self.fileData else { return }
+        if let error {
+            assertionFailure(error.localizedDescription)
+        }
+        
+        guard didChange else {
+            // update the document's fileModificationDate for a workaround (2014-03 by 1024jp)
+            // -> If not, an alert shows up when user saves the file.
+            guard let modificationDate else { return }
+            DispatchQueue.main.async { [weak self] in
+                if self?.fileModificationDate?.compare(modificationDate) == .orderedAscending {
+                    self?.fileModificationDate = modificationDate
+                }
+            }
+            return
+        }
         
         // notify about external file update
         Task {
@@ -732,7 +745,6 @@ final class Document: NSDocument, AdditionalDocumentPreparing, EncodingChanging 
     }
     
     
-    /// Apply the current states to menu items.
     override func validateMenuItem(_ menuItem: NSMenuItem) -> Bool {
         
         switch menuItem.action {