+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Fixes
 
+- Fix an issue on CotEditor 4.7.0 that documents did not update when the file was modified by a command-line process.
 - Fix an issue on CotEditor 4.7.0 that some actions, such as scripts, could not performed via the Quick Action bar.
 - Fix an issue on CotEditor 4.7.0 that last letters of some script name with a period could be wrongly interpreted as shortcut keys.
 - Fix an issue on CotEditor 4.7.0 that the Quick Actions bar dropped to draw the Shift key in some shortcuts.